% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_OPG_basic.R
\name{theme_OPG_basic}
\alias{theme_OPG_basic}
\title{A basic universal company theme}
\usage{
theme_OPG_basic()
}
\description{
Imports a ggplot 2 thgeme object to help you theme plot within company guidelines
}
